backend/requirements/base.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
django-cors-headers==4.9.0
django-privacy==0.2
psycopg2-binary==2.9.10
redis==6.4.0
celery==5.5.3
python-decouple==3.8
python-jose==3.5.0
pillow==11.3.0
cryptography==46.0.1
httpx==0.28.1
python-dateutil==2.9.0.post0
whitenoise==6.11.0
geoalchemy2==0.18.0
passlib==1.7.4
middleware==1.2.3
pydantic-settings==2.10.1